Logging system now takes variable parameters printf style

+++ core/src/input/input_manager.h
@@ -13,7 +13,7 @@ namespace core {	namespace input {
 		InputManager(state::StateManager manager);
 
 		void setCommand(int key, Command *command);
-		void cursorPositionCallback(int xpos, int ypos);
+		void cursorPositionCallback(double xpos, double ypos);
 		void mouseButtonCallback(int button, int action, int mods);
 		void mouseScrollCallback(double xoff, double yoff);
 		void keyCallback(int key, int scancode, int action, int mods);
